Move AuditHelper::getAuditTableColumns() and AuditHelper::getAuditTableIndices() to a dedicated helper
